+++ src/components/Button/Button.tsx
@@ -1,12 +1,5 @@
 import clsx from "clsx";
-import { ReactNode } from "react";
-
-type ButtonProps = {
-  children: ReactNode | string;
-  isDisabled?: boolean;
-  onClick: () => void;
-  className?: string;
-};
+import { ButtonProps } from "./Button.types";
 
 export function Button({
   children,
@@ -19,7 +12,7 @@ export function Button({
       role="button"
       tabIndex={0}
       className={clsx(
-        "flex py-2.5 px-5 shadow-sm text-sm font-medium transition-all bg-white rounded-lg border border-solid border-gray-400 hover:bg-gray-100 ",
+        "flex py-2.5 px-5 shadow-sm text-sm font-medium transition-all rounded-lg border border-solid border-gray-400 hover:bg-gray-100 ",
         isDisabled
           ? "cursor-auto opacity-40"
           : "cursor-pointer hover:opacity-80",
