+++ src/components/Button/Button.types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export type ButtonProps = {
+  children: ReactNode | string;
+  onClick: () => void;
+  isDisabled?: boolean;
+  className?: string;
+};
