+++ src/components/ListItem/ListItem.types.ts
@@ -1,0 +1,5 @@
+export type ListItemProps = {
+  text: string;
+  onClick: () => void;
+  isChecked?: boolean;
+};
