+++ src/components/Pagination/constants.ts
@@ -1,3 +1,3 @@
-import { RowsPerPageOptionsType } from "./Pagination";
+import { RowsPerPageOptionsType } from "./Pagination.types";
 
 export const rowsPerPageOptions: RowsPerPageOptionsType[] = [5, 10, 15, 25, 50];
